@@ -1,4 +1,4 @@
-class Arcana < ActiveRecord::Base
+class Arcana < ApplicationRecord
 
   default_scope do
     includes([
@@ -10,11 +10,11 @@ class Arcana < ActiveRecord::Base
   belongs_to :voice_actor
   belongs_to :illustrator
   belongs_to :first_skill,    class_name: 'Skill'
-  belongs_to :second_skill,   class_name: 'Skill'
-  belongs_to :third_skill,    class_name: 'Skill'
-  belongs_to :first_ability,  class_name: 'Ability'
-  belongs_to :second_ability, class_name: 'Ability'
-  belongs_to :weapon_ability, class_name: 'Ability'
+  belongs_to :second_skill,   class_name: 'Skill',   optional: true
+  belongs_to :third_skill,    class_name: 'Skill',   optional: true
+  belongs_to :first_ability,  class_name: 'Ability', optional: true
+  belongs_to :second_ability, class_name: 'Ability', optional: true
+  belongs_to :weapon_ability, class_name: 'Ability', optional: true
   belongs_to :chain_ability
 
   RARITYS = (1..5).freeze
